feat(accuracy.js, results.js): created new file
