+++ index.html
@@ -125,8 +125,8 @@
                   <!-- <i class="fas fa-users"></i> -->
                 </button>
                 <div id="toggle-pairing-filter" class="toggle-pairing-filter" style="display: none;">
-            <button> by Location</button>
-            <button> by Department</button>
+            <button> same Location</button>
+            <button> same Department</button>
                  </div>
               <button id="generate-pairing" class="generate-button">
                 Generate Pairing
