+++ results.js
@@ -1,4 +1,5 @@
 import * as generate_CSV from "./src/util/generate_csv.js"; 
+import * as Accuracy from "./src/util/accuracy.js";
 
 document.addEventListener("DOMContentLoaded", () => {
     let internPairs = []; 
@@ -69,7 +70,7 @@ document.addEventListener("DOMContentLoaded", () => {
         window_functionality_setup();
         edit_button_functionality_setup();
         add_button_functionality_setup();
-        calculatePairAccuracy();
+        Accuracy.calculatePairAccuracy(internPairs);
     };
 
     //event listener for download csv
@@ -435,46 +436,6 @@ document.addEventListener("DOMContentLoaded", () => {
     //Add All Previously Remove.
     //Previously Remove
 
-    function calculatePairAccuracy() {
-        let isPairedByLocation = JSON.parse(sessionStorage.getItem('isPairedByLocation'));
-        let isPairedByDepartment = JSON.parse(sessionStorage.getItem('isPairedByDepartment'));
-        let validPairs = 0;
-       
-        if (!isPairedByLocation && !isPairedByDepartment) {
-            console.log('Pairing accuracy: N/A');
-            return;
-        }
- 
- 
-        internPairs.forEach(pair => {
-            const [intern1, intern2] = pair;
-   
-            let isValid = true;
-           
-            // console.log('Checking pair:', intern1, intern2);
-   
-            if (isPairedByLocation) {
-                isValid = isValid && intern1.location === intern2.location;
-                // console.log(`Location check: ${isValid}`);
-            }
-   
-            if (isPairedByDepartment) {
-                isValid = isValid && intern1.department === intern2.department;
-                // console.log(`Department check: ${isValid}`);
-            }
-   
-            if (isValid) {
-                validPairs++;
-                // console.log('Valid pair found');
-            }
-        });
-   
-        const totalPairs = internPairs.length;
-        const accuracy = totalPairs > 0 ? (validPairs / totalPairs) * 100 : 0;
-   
-        console.log(`Pairing accuarcy: ${accuracy.toFixed(2)}%`);
-    }
- 
 });
